now only loggedin users can delete their own threads

+++ controllers/threadController.js
@@ -79,11 +79,24 @@ const updateThread = async (req, res) => {
 const deleteThread = async (req, res) => {
   try {
     const { threadId } = req.params;
-    const deletedThread = await Thread.findByIdAndDelete(threadId);
-    if (!deletedThread) {
+    const { userId } = req.session;
+
+    const thread = await Thread.findById(threadId);
+
+    if (!thread) {
       console.log("Thread not found");
       return res.redirect("/not-found");
     }
+
+    // Check if the current user is the author of the thread
+    if (thread.author.toString() !== userId) {
+      console.log("You are not authorized to delete this thread");
+      return res.redirect("/not-authorized");
+    }
+
+    // Delete the thread
+    await Thread.findByIdAndDelete(threadId);
+
     console.log("Thread deleted successfully");
     return res.redirect("/");
   } catch (error) {
@@ -91,7 +104,6 @@ const deleteThread = async (req, res) => {
     res.redirect("/error");
   }
 };
-
 module.exports = {
   createThread,
   getThread,
